audiodescription/requirements.txt
pyautogui
Pillow
opencv-python
numpy
openai
pydub
pygame
adb